requirements.txt
langchain==0.1.20
langchain_chroma==0.1.0
langchain_community==0.0.38
langchain_core==0.1.52
langchain_text_splitters==0.0.1
langsmith==0.1.57
openai_whisper==20231117
openpyxl==3.1.2
pandas==2.2.2
python-dotenv==1.0.1
openai
openai-whisper
pip install git+https://github.com/openai/whisper.git
sudo apt update && sudo apt install ffmpeg